feat: moved models into model.py
