+++ app.py
@@ -1,10 +1,9 @@
 import os
 
-from datetime import datetime
-
 from flask import Flask, flash, render_template, redirect, request
 from flask_session import Session
-from flask_sqlalchemy import SQLAlchemy
+
+from models import db, Category, Skill, User, UserSkill
 
 # Create Flask app
 app = Flask(__name__)
@@ -21,79 +20,8 @@ app.config["SESSION_TYPE"] = "filesystem"
 
 # Initialize Session
 Session(app)
-# Initialize SQLAlchemy
-db = SQLAlchemy(app)
-
-
-class Category(db.Model):
-    __tablename__ = "categories"
-
-    # Fields
-    id = db.Column(db.Integer, primary_key=True, autoincrement=True)
-    name = db.Column(db.Text, nullable=False, unique=True)
-
-    def __repr__(self):
-        return f"<Category {self.name}>"
-
-
-class Skill(db.Model):
-    __tablename__ = "skills"
-
-    # Fields
-    id = db.Column(db.Integer, primary_key=True, autoincrement=True)
-    name = db.Column(db.Text, nullable=False, unique=True)
-    category_id = db.Column(db.Integer, db.ForeignKey("categories.id"), nullable=False)
-
-    # Relationships
-    category = db.relationship("Category", backref="skills")
-    user_skills = db.relationship(
-        "UserSkill", back_populates="skill", cascade="all, delete-orphan"
-    )
-
-    def __repr__(self):
-        return f"<Skill {self.name}>"
-
-
-class User(db.Model):
-    __tablename__ = "users"
-
-    # Fields
-    id = db.Column(db.Integer, primary_key=True, autoincrement=True)
-    username = db.Column(db.Text, nullable=False, unique=True)
-    email = db.Column(db.Text, nullable=False, unique=True)
-    hash = db.Column(db.Text, nullable=False)
-
-    # Relationships
-    user_skills = db.relationship(
-        "UserSkill", back_populates="user", cascade="all, delete-orphan"
-    )
-
-    def __repr__(self):
-        return f"<User {self.username}>"
-
-
-class UserSkill(db.Model):
-    __tablename__ = "user_skills"
-
-    # Fields
-    id = db.Column(db.Integer, primary_key=True, autoincrement=True)
-    user_id = db.Column(db.Integer, db.ForeignKey("users.id"), nullable=False)
-    skill_id = db.Column(db.Integer, db.ForeignKey("skills.id"), nullable=False)
-    date_added = db.Column(db.Date, nullable=False, default=datetime.now())
-    source = db.Column(db.Text)
-    application = db.Column(db.Text)
-    mastery_level = db.Column(
-        db.Integer, db.CheckConstraint("mastery_level >= 1 AND mastery_level <= 10")
-    )
-    category = db.Column(db.Text)
-
-    # Relationships
-    user = db.relationship("User", back_populates="user_skills")
-    skill = db.relationship("Skill", back_populates="user_skills")
-
-    def __repr__(self):
-        return f"<UserSkill User={self.user_id} Skill={self.skill_id} Mastery={self.mastery_level}>"
-
+# Initialize SQLAlchemy with app
+db.init_app(app)
 
 @app.route("/")
 def index():
@@ -169,6 +97,7 @@ def get_skills(id=None, name=None):
         return Skill.query.filter_by(id=id).first()
     elif name:
         return Skill.query.filter_by(name=name).first()
+    return Skill.query.all()
 
 
 def get_categories(id=None, name=None):
@@ -178,4 +107,3 @@ def get_categories(id=None, name=None):
     elif name:
         return Category.query.filter_by(name=name).first()
     return Category.query.all()
-    return Skill.query.all()
